+++ plugins/rbac-backend/src/service/policy-builder.ts
@@ -3,6 +3,7 @@ import {
   PluginEndpointDiscovery,
   resolvePackagePath,
 } from '@backstage/backend-common';
+import { CatalogClient } from '@backstage/catalog-client';
 import { Config } from '@backstage/config';
 import { IdentityApi } from '@backstage/plugin-auth-node';
 import { RouterOptions } from '@backstage/plugin-permission-backend';
@@ -16,6 +17,7 @@ import { CasbinDBAdapterFactory } from './casbin-adapter-factory';
 import { MODEL } from './permission-model';
 import { RBACPermissionPolicy } from './permission-policy';
 import { PolicesServer } from './policies-rest-api';
+import { BackstageRoleManager } from './role-manager';
 
 export class PolicyBuilder {
   public static async build(env: {
@@ -48,6 +50,12 @@ export class PolicyBuilder {
     const enf = await newEnforcer(newModelFromString(MODEL), adapter);
     await enf.loadPolicy();
     enf.enableAutoSave(true);
+
+    const catalogClient = new CatalogClient({ discoveryApi: env.discovery });
+    const rm = new BackstageRoleManager(catalogClient, env.logger);
+    enf.setRoleManager(rm);
+    enf.enableAutoBuildRoleLinks(false);
+    await enf.buildRoleLinks();
 
     const options: RouterOptions = {
       config: env.config,